# artifacts/api-server/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Order Automation System API",
    "description": "API for automated payment and order management system. Includes product management, order processing, webhook handling, and admin reporting.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    },
    {
      "url": "https://order-automation-api.onrender.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create new order",
        "description": "Create a new order with product and customer information",
        "operationId": "createOrder",
        "security": [
          {
            "X-CSRF-Token": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID",
        "description": "Retrieve detailed information about a specific order",
        "operationId": "getOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/ship": {
      "patch": {
        "tags": ["Orders"],
        "summary": "Mark order as shipped",
        "description": "Admin endpoint to mark an order as shipped",
        "operationId": "markOrderShipped",
        "security": [
          {
            "X-Telegram-Chat-Id": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Order marked as shipped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/reports": {
      "get": {
        "tags": ["Reports"],
        "summary": "List payment reports",
        "description": "Admin endpoint to list payment reports with pagination",
        "operationId": "listReports",
        "security": [
          {
            "X-Telegram-Chat-Id": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of payment reports",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reports/order/:orderId": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get payment reports for a specific order",
        "description": "Admin endpoint to get payment reports for a specific order",
        "operationId": "getOrderReports",
        "security": [
          {
            "X-Telegram-Chat-Id": []
          }
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Payment reports for the order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportsResponse"
                }
              }
            }
          },
          "404": {
            "description": "No payment reports found for this order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/reports/orders": {
      "get": {
        "tags": ["Reports"],
        "summary": "List orders with details",
        "description": "Admin endpoint to list orders with full details",
        "operationId": "listOrdersReport",
        "security": [
          {
            "X-Telegram-Chat-Id": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["pending", "paid", "shipped", "cancelled"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersReportResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "SePay webhook token (format: Apikey <token>)"
      },
      "X-CSRF-Token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF protection token"
      },
      "X-Telegram-Chat-Id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Telegram-Chat-Id",
        "description": "Telegram admin chat ID for authentication"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number"
          },
          "image_url": {
            "type": "string",
            "nullable": true
          },
          "stock_quantity": {
            "type": "integer"
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateOrderRequest": {
        "type": "object",
        "required": ["product_id", "quantity", "customer_name"],
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "customer_name": {
            "type": "string",
            "maxLength": 255
          },
          "customer_email": {
            "type": "string",
            "format": "email"
          },
          "customer_phone": {
            "type": "string",
            "pattern": "^(0|\\+84)\\d{9,10}$"
          },
          "customer_note": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "CreateOrderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "order": {
                "$ref": "#/components/schemas/Order"
              },
              "product": {
                "$ref": "#/components/schemas/Product"
              },
              "qr_code_url": {
                "type": "string"
              }
            }
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "order_code": {
            "type": "string"
          },
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "quantity": {
            "type": "integer"
          },
          "total_amount": {
            "type": "number"
          },
          "customer_name": {
            "type": "string",
            "nullable": true
          },
          "customer_email": {
            "type": "string",
            "nullable": true
          },
          "customer_phone": {
            "type": "string",
            "nullable": true
          },
          "customer_note": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["pending", "paid", "shipped", "cancelled"]
          },
          "payment_content": {
            "type": "string",
            "nullable": true
          },
          "paid_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shipped_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelled_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "telegram_message_id": {
            "type": "integer",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "OrderDetail": {
        "type": "object",
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid"
          },
          "order_code": {
            "type": "string"
          },
          "order_status": {
            "type": "string",
            "enum": ["pending", "paid", "shipped", "cancelled"]
          },
          "total_amount": {
            "type": "number"
          },
          "quantity": {
            "type": "integer"
          },
          "customer_name": {
            "type": "string",
            "nullable": true
          },
          "customer_email": {
            "type": "string",
            "nullable": true
          },
          "customer_phone": {
            "type": "string",
            "nullable": true
          },
          "payment_content": {
            "type": "string",
            "nullable": true
          },
          "paid_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "shipped_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelled_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "order_created_at": {
            "type": "string",
            "format": "date-time"
          },
          "product_id": {
            "type": "string",
            "format": "uuid"
          },
          "product_name": {
            "type": "string"
          },
          "product_price": {
            "type": "number"
          },
          "transaction_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "sepay_transaction_id": {
            "type": "integer",
            "nullable": true
          },
          "paid_amount": {
            "type": "number",
            "nullable": true
          },
          "transaction_date": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "bank_brand_name": {
            "type": "string",
            "nullable": true
          },
          "transaction_status": {
            "type": "string",
            "enum": ["matched", "unmatched", "duplicate"],
            "nullable": true
          },
          "report_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "bill_image_url": {
            "type": "string",
            "nullable": true
          },
          "report_created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Order"
          }
        }
      },
      "OrderDetailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/OrderDetail"
          }
        }
      },
...
      "ReportsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "order_id": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "transaction_id": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "bill_image_url": {
                  "type": "string"
                },
                "cloudinary_public_id": {
                  "type": "string"
                },
                "uploaded_by_telegram_id": {
                  "type": "integer",
                  "nullable": true
                },
                "note": {
                  "type": "string",
                  "nullable": true
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "OrdersReportResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDetail"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true
          }
        }
      }
    }
  }
}